Refactor patches as helpers to use standalone where required
