+++ modules/transport.js
@@ -21,7 +21,7 @@ let Broadcaster = require('../logic/broadcaster.js');
 const failureCodes = require('../api/ws/rpc/failure_codes');
 const PeerUpdateError = require('../api/ws/rpc/failure_codes').PeerUpdateError;
 const Rules = require('../api/ws/workers/rules');
-const Patches = require('../modules/patches');
+const patches = require('../helpers/patches');
 // eslint-disable-next-line prefer-const
 let wsRPC = require('../api/ws/rpc/ws_rpc').wsRPC;
 
@@ -317,7 +317,6 @@ Transport.prototype.onBind = function(scope) {
 		peers: scope.peers,
 		system: scope.system,
 		transactions: scope.transactions,
-		patches: new Patches(library.logger),
 	};
 
 	definitions = scope.swagger.definitions;
@@ -397,9 +396,10 @@ Transport.prototype.broadcastHeaders = cb => {
 	async.each(
 		peers,
 		(peer, eachCb) => {
-			const peerObject = modules.patches.systemHeaders.versionForPreRelease(
+			const peerObject = patches.systemHeaders.versionForPreRelease(
 				peer.version,
-				library.logic.peers.me()
+				library.logic.peers.me(),
+				library.logger
 			);
 			peer.rpc.updateMyself(peerObject, err => {
 				if (err) {
